+++ design_system/src/main/res/values-ar/strings.xml
@@ -25,8 +25,6 @@
     <string name="reviews">المراجعات</string>
     <string name="rating">التقييم</string>
     <string name="date">التاريخ</string>
-
-
     <!-- Tab Labels -->
     <string name="movies">أفلام</string>
     <string name="tv_shows">مسلسلات</string>
